+++ src/nn/activations.rs
@@ -19,12 +19,6 @@ macro_rules! activation_impls {
             fn reset_params<R: Rng>(&mut self, _: &mut R) {}
         }
 
-        #[cfg(feature = "numpy")]
-        impl SaveToNpz for $struct_name {}
-
-        #[cfg(feature = "numpy")]
-        impl LoadFromNpz for $struct_name {}
-
         impl<T: Tensor<Dtype = f32>> Module<T> for $struct_name {
             type Output = T;
             fn forward(&self, input: T) -> Self::Output {
@@ -68,12 +62,6 @@ impl ResetParams for Softmax {
     /// Does nothing.
     fn reset_params<R: Rng>(&mut self, _: &mut R) {}
 }
-
-#[cfg(feature = "numpy")]
-impl SaveToNpz for Softmax {}
-
-#[cfg(feature = "numpy")]
-impl LoadFromNpz for Softmax {}
 
 impl<T> Module<T> for Softmax
 where
