+++ src/nn/conv.rs
@@ -3,11 +3,6 @@ use crate::prelude::*;
 use rand::Rng;
 use rand_distr::Uniform;
 
-#[cfg(feature = "numpy")]
-use std::io::{Read, Seek, Write};
-#[cfg(feature = "numpy")]
-use zip::{result::ZipResult, ZipArchive, ZipWriter};
-
 /// **Requires Nightly** Performs 2d convolutions on 3d and 4d images.
 ///
 /// **Pytorch Equivalent**: `torch.nn.Conv2d`
@@ -57,32 +52,6 @@ impl<const I: usize, const O: usize, const K: usize, const S: usize, const P: us
         let dist = Uniform::new(-bound, bound);
         self.weight.randomize(rng, &dist);
         self.bias.randomize(rng, &dist);
-    }
-}
-
-#[cfg(feature = "numpy")]
-impl<const I: usize, const O: usize, const K: usize, const S: usize, const P: usize> SaveToNpz
-    for Conv2D<I, O, K, S, P>
-{
-    /// Saves [Self::weight] to `{pre}weight.npy` and [Self::bias] to `{pre}bias.npy`
-    /// using [npz_fwrite()].
-    fn write<W: Write + Seek>(&self, pre: &str, w: &mut ZipWriter<W>) -> ZipResult<()> {
-        npz_fwrite(w, format!("{pre}weight.npy"), self.weight.data())?;
-        npz_fwrite(w, format!("{pre}bias.npy"), self.bias.data())?;
-        Ok(())
-    }
-}
-
-#[cfg(feature = "numpy")]
-impl<const I: usize, const O: usize, const K: usize, const S: usize, const P: usize> LoadFromNpz
-    for Conv2D<I, O, K, S, P>
-{
-    /// Reads [Self::weight] from `{pre}weight.npy` and [Self::bias] from `{pre}bias.npy`
-    /// using [npz_fread()].
-    fn read<R: Read + Seek>(&mut self, pre: &str, r: &mut ZipArchive<R>) -> Result<(), NpzError> {
-        npz_fread(r, format!("{pre}weight.npy"), self.weight.mut_data())?;
-        npz_fread(r, format!("{pre}bias.npy"), self.bias.mut_data())?;
-        Ok(())
     }
 }
 
@@ -188,54 +157,6 @@ mod tests {
 
         type Img = Tensor3D<1, 10, 10>;
         let _: Tensor3D<1, 8, 8> = <(A, B, C)>::default().forward_mut(Img::zeros());
-    }
-
-    #[cfg(feature = "numpy")]
-    mod numpy_tests {
-        use super::*;
-        use std::fs::File;
-        use tempfile::NamedTempFile;
-
-        #[test]
-        fn test_save_conv2d() {
-            let model: Conv2D<2, 4, 3> = Default::default();
-            let file = NamedTempFile::new().expect("failed to create tempfile");
-            model
-                .save(file.path().to_str().unwrap())
-                .expect("failed to save model");
-            let f = File::open(file.path()).expect("failed to open resulting file");
-            let mut zip = ZipArchive::new(f).expect("failed to create zip archive from file");
-            {
-                let weight_file = zip
-                    .by_name("weight.npy")
-                    .expect("failed to find weight.npy file");
-                assert!(weight_file.size() > 0);
-            }
-            {
-                let bias_file = zip
-                    .by_name("bias.npy")
-                    .expect("failed to find bias.npy file");
-                assert!(bias_file.size() > 0);
-            }
-        }
-
-        #[test]
-        fn test_load_conv() {
-            let mut rng = thread_rng();
-            let mut saved_model: Conv2D<2, 4, 3> = Default::default();
-            saved_model.reset_params(&mut rng);
-
-            let file = NamedTempFile::new().expect("failed to create tempfile");
-            assert!(saved_model.save(file.path().to_str().unwrap()).is_ok());
-
-            let mut loaded_model: Conv2D<2, 4, 3> = Default::default();
-            assert!(loaded_model.weight.data() != saved_model.weight.data());
-            assert!(loaded_model.bias.data() != saved_model.bias.data());
-
-            assert!(loaded_model.load(file.path().to_str().unwrap()).is_ok());
-            assert_eq!(loaded_model.weight.data(), saved_model.weight.data());
-            assert_eq!(loaded_model.bias.data(), saved_model.bias.data());
-        }
     }
 
     #[test]
