+++ src/nn/flatten.rs
@@ -23,11 +23,6 @@ impl CanUpdateWithGradients for FlattenImage {
     fn update<G: GradientProvider>(&mut self, _: &mut G, _: &mut UnusedTensors) {}
 }
 
-#[cfg(feature = "numpy")]
-impl SaveToNpz for FlattenImage {}
-#[cfg(feature = "numpy")]
-impl LoadFromNpz for FlattenImage {}
-
 impl<const M: usize, const N: usize, const O: usize, H: Tape> Module<Tensor3D<M, N, O, H>>
     for FlattenImage
 where
