+++ src/nn/impl_module_for_tuples.rs
@@ -1,10 +1,6 @@
 use crate::gradients::{CanUpdateWithGradients, GradientProvider, UnusedTensors};
 use crate::prelude::*;
 use rand::prelude::Rng;
-#[cfg(feature = "numpy")]
-use std::io::{Read, Seek, Write};
-#[cfg(feature = "numpy")]
-use zip::{result::ZipResult, ZipArchive, ZipWriter};
 
 macro_rules! tuple_impls {
     ([$($name:ident),+] [$($idx:tt),+], $last:ident, [$($rev_tail:ident),+]) => {
@@ -17,32 +13,6 @@ macro_rules! tuple_impls {
         impl<$($name: ResetParams),+> ResetParams for ($($name,)+) {
             fn reset_params<R: Rng>(&mut self, rng: &mut R) {
                 $(self.$idx.reset_params(rng));+
-            }
-        }
-
-        #[cfg(feature = "numpy")]
-        impl<$($name: SaveToNpz),+> SaveToNpz for ($($name,)+) {
-            /// Calls `SaveToNpz::write(self.<idx>, ...)` on each part of the tuple. See [SaveToNpz].
-            ///
-            /// E.g. for a two tuple (A, B) with `base == ""`, this will call:
-            /// 1. `self.0.write("0.", w)`
-            /// 2. `self.1.write("1.", w)`
-            fn write<W: Write + Seek>(&self, base: &str, w: &mut ZipWriter<W>) -> ZipResult<()> {
-                $(self.$idx.write(&format!("{}{}.", base, $idx), w)?;)+
-                Ok(())
-            }
-        }
-
-        #[cfg(feature = "numpy")]
-        impl<$($name: LoadFromNpz),+> LoadFromNpz for ($($name,)+) {
-            /// Calls `LoadFromNpz::read(self.<idx>, ...)` on each part of the tuple. See [LoadFromNpz].
-            ///
-            /// E.g. for a two tuple (A, B) with `base == ""`, this will call:
-            /// 1. `self.0.read("0.", r)`
-            /// 2. `self.1.read("1.", r)`
-            fn read<R: Read + Seek>(&mut self, base: &str, r: &mut ZipArchive<R>) -> Result<(), NpzError> {
-                $(self.$idx.read(&format!("{}{}.", base, $idx), r)?;)+
-                Ok(())
             }
         }
 
@@ -158,79 +128,6 @@ mod tests {
         assert!(model.0.bias.data() != m0.0.bias.data());
         assert!(model.1.weight.data() != m0.1.weight.data());
         assert!(model.1.bias.data() != m0.1.bias.data());
-    }
-
-    #[cfg(feature = "numpy")]
-    mod numpy_tests {
-        use super::*;
-        use std::fs::File;
-        use tempfile::NamedTempFile;
-
-        #[test]
-        fn test_save_tuple() {
-            let model: (
-                Linear<1, 2>,
-                ReLU,
-                Linear<2, 3>,
-                (Dropout, Linear<1, 2>, Linear<3, 4>),
-            ) = Default::default();
-            let file = NamedTempFile::new().expect("failed to create tempfile");
-            model
-                .save(file.path().to_str().unwrap())
-                .expect("failed to save model");
-            let f = File::open(file.path()).expect("failed to open resulting file");
-            let zip = ZipArchive::new(f).expect("failed to create zip archive from file");
-            let mut names = zip.file_names().collect::<Vec<&str>>();
-            names.sort_unstable();
-            assert_eq!(
-                &names,
-                &[
-                    "0.bias.npy",
-                    "0.weight.npy",
-                    "2.bias.npy",
-                    "2.weight.npy",
-                    "3.1.bias.npy",
-                    "3.1.weight.npy",
-                    "3.2.bias.npy",
-                    "3.2.weight.npy",
-                ]
-            );
-        }
-
-        #[test]
-        fn test_load_tuple() {
-            type Model = (
-                Linear<1, 2>,
-                ReLU,
-                Linear<2, 3>,
-                (Dropout, Linear<1, 2>, Linear<3, 4>),
-            );
-
-            let mut rng = StdRng::seed_from_u64(0);
-            let mut saved_model: Model = Default::default();
-            saved_model.reset_params(&mut rng);
-
-            let file = NamedTempFile::new().expect("failed to create tempfile");
-            assert!(saved_model.save(file.path().to_str().unwrap()).is_ok());
-
-            let mut loaded_model: Model = Default::default();
-            assert!(loaded_model.load(file.path().to_str().unwrap()).is_ok());
-            assert_eq!(loaded_model.0.weight.data(), saved_model.0.weight.data());
-            assert_eq!(loaded_model.0.bias.data(), saved_model.0.bias.data());
-            assert_eq!(loaded_model.2.weight.data(), saved_model.2.weight.data());
-            assert_eq!(loaded_model.2.bias.data(), saved_model.2.bias.data());
-            assert_eq!(
-                loaded_model.3 .1.weight.data(),
-                saved_model.3 .1.weight.data()
-            );
-            assert_eq!(loaded_model.3 .1.bias.data(), saved_model.3 .1.bias.data());
-
-            assert_eq!(
-                loaded_model.3 .2.weight.data(),
-                saved_model.3 .2.weight.data()
-            );
-            assert_eq!(loaded_model.3 .2.bias.data(), saved_model.3 .2.bias.data());
-        }
     }
 
     /// A struct to test the forward method of tuples. This sets the `I`th valuein a 1d tensors of size `N` to 1.0.
