+++ src/nn/pool2d.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "numpy")]
-use super::{LoadFromNpz, SaveToNpz};
 use super::{Module, ModuleMut, ResetParams};
 use crate::gradients::*;
 use crate::tensor::*;
@@ -46,11 +44,6 @@ macro_rules! impl_pools {
         impl<const K: usize, const S: usize, const P: usize> ResetParams for $PoolTy<K, S, P> {
             fn reset_params<R: Rng>(&mut self, _: &mut R) {}
         }
-
-        #[cfg(feature = "numpy")]
-        impl<const K: usize, const S: usize, const P: usize> SaveToNpz for $PoolTy<K, S, P> {}
-        #[cfg(feature = "numpy")]
-        impl<const K: usize, const S: usize, const P: usize> LoadFromNpz for $PoolTy<K, S, P> {}
 
         impl<
                 const K: usize,
