+++ src/nn/pool_global.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "numpy")]
-use super::{LoadFromNpz, SaveToNpz};
 use super::{Module, ModuleMut, ResetParams};
 use crate::gradients::*;
 use crate::tensor::*;
@@ -66,12 +64,6 @@ macro_rules! impl_pools {
         impl CanUpdateWithGradients for $PoolTy {
             fn update<G: GradientProvider>(&mut self, _: &mut G, _: &mut UnusedTensors) {}
         }
-
-        #[cfg(feature = "numpy")]
-        impl SaveToNpz for $PoolTy {}
-
-        #[cfg(feature = "numpy")]
-        impl LoadFromNpz for $PoolTy {}
 
         impl<const C: usize, const L: usize, T: Tape> Module<Tensor2D<C, L, T>> for $PoolTy {
             type Output = Tensor1D<C, T>;
