+++ src/nn/residual.rs
@@ -1,9 +1,5 @@
 use crate::gradients::{CanUpdateWithGradients, GradientProvider, UnusedTensors};
 use crate::prelude::*;
-#[cfg(feature = "numpy")]
-use std::io::{Read, Seek, Write};
-#[cfg(feature = "numpy")]
-use zip::{result::ZipResult, ZipArchive, ZipWriter};
 
 /// A residual connection around `F`: `F(x) + x`,
 /// as introduced in [Deep Residual Learning for Image Recognition](https://arxiv.org/abs/1512.03385).
@@ -52,22 +48,6 @@ impl<T: Tensor<Dtype = f32>, F: ModuleMut<T, Output = T>> ModuleMut<T> for Resid
     }
 }
 
-#[cfg(feature = "numpy")]
-impl<F: SaveToNpz> SaveToNpz for Residual<F> {
-    /// Pass through to `F`'s [SaveToNpz].
-    fn write<W: Write + Seek>(&self, p: &str, w: &mut ZipWriter<W>) -> ZipResult<()> {
-        self.0.write(p, w)
-    }
-}
-
-#[cfg(feature = "numpy")]
-impl<F: LoadFromNpz> LoadFromNpz for Residual<F> {
-    /// Pass through to `F`'s [LoadFromNpz].
-    fn read<R: Read + Seek>(&mut self, p: &str, r: &mut ZipArchive<R>) -> Result<(), NpzError> {
-        self.0.read(p, r)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -103,29 +83,5 @@ mod tests {
         assert_close(g.ref_gradient(&model.0.weight), &[[0.475242, -0.075136]; 2]);
         assert_close(g.ref_gradient(&model.0.bias), &[0.5; 2]);
         assert_close(g.ref_gradient(&x), &[[0.18806472, 0.21419683]; 4]);
-    }
-
-    #[cfg(feature = "numpy")]
-    mod numpy_tests {
-        use super::*;
-        use tempfile::NamedTempFile;
-
-        #[test]
-        fn test_save_load_residual() {
-            let mut rng = StdRng::seed_from_u64(0);
-            let mut saved_model: Residual<Linear<5, 3>> = Default::default();
-            saved_model.reset_params(&mut rng);
-
-            let file = NamedTempFile::new().expect("failed to create tempfile");
-            assert!(saved_model.save(file.path().to_str().unwrap()).is_ok());
-
-            let mut loaded_model: Residual<Linear<5, 3>> = Default::default();
-            assert_ne!(loaded_model.0.weight.data(), saved_model.0.weight.data());
-            assert_ne!(loaded_model.0.bias.data(), saved_model.0.bias.data());
-
-            assert!(loaded_model.load(file.path().to_str().unwrap()).is_ok());
-            assert_eq!(loaded_model.0.weight.data(), saved_model.0.weight.data());
-            assert_eq!(loaded_model.0.bias.data(), saved_model.0.bias.data());
-        }
     }
 }
